Create class PdoConnection
